eliminated util folder and moved necessary files to main directory, fixed a typo in unit test
