+++ corgidrp/detector.py
@@ -5,158 +5,6 @@ import corgidrp.data as data
 from scipy import interpolate
 from scipy.ndimage import median_filter
 
-from astropy.time import Time
-import os
-from pathlib import Path
-import yaml
-
-class ReadMetadataException(Exception):
-    """Exception class for read_metadata module."""
-
-# Set up to allow the metadata.yaml in the repo be the default
-here = Path(os.path.dirname(os.path.abspath(__file__)))
-meta_path = Path(here, 'util', 'metadata.yaml')
-
-class Metadata(object):
-    """Create masks for different regions of the Roman CGI EXCAM detector.
-
-    Parameters
-    ----------
-    meta_path : str
-        Full path of metadta yaml.
-
-    Attributes
-    ----------
-    data : dict
-        Data from metadata file.
-    geom : SimpleNamespace
-        Geometry specific data.
-
-    B Nemati and S Miller - UAH - 03-Aug-2018
-
-    """
-
-    def __init__(self, meta_path=meta_path, obstype='SCI'):
-        self.meta_path = meta_path
-
-        self.data = self.get_data()
-        self.obstype = obstype
-        self.frame_rows = self.data[obstype]['frame_rows']
-        self.frame_cols = self.data[obstype]['frame_cols']
-        self.geom = self.data[obstype]['geom']
-
-        # Get imaging area geometry
-        self.rows_im, self.cols_im, self.r0c0_im = self._imaging_area_geom()
-        # Make some zeros frames for initial creation of arrays
-        self.imaging_area_zeros = np.zeros((self.rows_im, self.cols_im))
-        self.full_frame_zeros = np.zeros((self.frame_rows, self.frame_cols))
-
-    def get_data(self):
-        """Read yaml data into dictionary.
-
-        Returns:
-            data: data from the .yaml file
-        """
-        with open(self.meta_path, 'r') as stream:
-            data = yaml.safe_load(stream)
-        return data
-
-    def mask(self, key):
-        full_frame_m = self.full_frame_zeros.copy()
-
-        rows, cols, r0c0 = self._unpack_geom(key)
-        full_frame_m[r0c0[0]:r0c0[0]+rows, r0c0[1]:r0c0[1]+cols] = 1
-        return full_frame_m.astype(bool)
-
-    def slice_section(self, frame, key):
-        """Slice 2d section out of frame.
-
-        Args:
-            frame: array_like
-            Full frame consistent with size given in frame_rows, frame_cols.
-            key: str
-            Keyword referencing section to be sliced; must exist in geom.
-
-        Returns:
-            section: array_like
-            Desired slice of the frame
-
-        """
-        rows, cols, r0c0 = self._unpack_geom(key)
-
-        section = frame[r0c0[0]:r0c0[0]+rows, r0c0[1]:r0c0[1]+cols]
-        if section.size == 0:
-            raise ReadMetadataException('Corners invalid')
-        return section
-
-    def _unpack_geom(self, key):
-        """Safely check format of geom sub-dictionary and return values.
-
-        Args:
-            key: str
-            Desired section
-
-        Returns:
-            rows: int
-            Number of rows of frame
-            cols : int
-            Number of columns of frame
-            r0c0: tuple
-            Tuple of (row position, column position) of corner closest to (0,0)
-        """
-        coords = self.geom[key]
-        rows = coords['rows']
-        cols = coords['cols']
-        r0c0 = coords['r0c0']
-
-        return rows, cols, r0c0
-
-    def _imaging_area_geom(self):
-        """Return geometry of imaging area (including shielded pixels)
-        in reference to full frame.  Different from normal image area.
-
-        Returns:
-            rows: int
-            Number of rows of imaging area
-            cols : int
-            Number of columns of imaging area
-            r0c0: tuple
-            Tuple of (row position, column position) of corner closest to (0,0)
-        """
-        _, cols_pre, _ = self._unpack_geom('prescan')
-        _, cols_serial_ovr, _ = self._unpack_geom('serial_overscan')
-        rows_parallel_ovr, _, _ = self._unpack_geom('parallel_overscan')
-        #_, _, r0c0_image = self._unpack_geom('image')
-        fluxmap_rows, _, r0c0_image = self._unpack_geom('image')
-
-        rows_im = self.frame_rows - rows_parallel_ovr
-        cols_im = self.frame_cols - cols_pre - cols_serial_ovr
-        r0c0_im = r0c0_image.copy()
-        r0c0_im[0] = r0c0_im[0] - (rows_im - fluxmap_rows)
-
-        return rows_im, cols_im, r0c0_im
-
-    def imaging_slice(self, frame):
-        """Select only the real counts from full frame and exclude virtual.
-        Includes shielded pixels.
-
-        Use this to transform mask and embed from acting on the full frame to
-        acting on only the image frame.
-
-        Args:
-            frame: array_like
-            Input frame
-
-        Returns:
-            sl: array_like
-            Imaging slice
-
-        """
-        rows, cols, r0c0 = self._imaging_area_geom()
-        sl = frame[r0c0[0]:r0c0[0]+rows, r0c0[1]:r0c0[1]+cols]
-        return sl
-
-
 def create_dark_calib(dark_dataset):
     """
     Turn this dataset of image frames that were taken to measure
