--- corgidrp/util/metadata.yaml
@@ -1,174 +1,0 @@
-#############
-# This file is the default file used by gsw_process.Process and is valid
-# for SCIENCE/ACQUIRE and TRAP PUMPING sequences.  The SCI format frame
-# is broken up from the ENG format frame.  For example, prescan specifies
-# the SCI prescan section of the frame, and this along with eng_prescan
-# specifies the entire prescan region of the ENG format frame for these
-# sequences.
-
-# This file also specifies the ENG format frame for
-# SCIENCE/ACQUIRE and TRAP PUMPING sequences.
-#############
-
-# Any regions not specified are assumed to be shielded or unused pixels.
-#
-# Geometry regions are defined as:
-#   rows: number of region rows
-#   cols: number of region columns
-#   r0c0: coordinates (row, col) of the corner closest to (0, 0) of the array
-#
-# For example, in a 5x5 array, the following geometry would look as follows:
-#   rows: 2
-#   cols: 3
-#   r0c0:
-#     - 2
-#     - 1
-#
-# [
-#     [0, 0, 0, 0, 0],
-#     [0, 0, 0, 0, 0],
-#     [0, 1, 1, 1, 0],
-#     [0, 1, 1, 1, 0],
-#     [0, 0, 0, 0, 0]
-# ]
-#
-# where (0, 0) is defined as the top left of this array.
-
-# total ENG frame:
-eng_frame_rows: 2200
-eng_frame_cols: 2200
-
-# SCI frame:
-SCI:
-  frame_rows: 1200
-  frame_cols: 2200
-  geom:
-    image:
-      rows: 1024
-      cols: 1024
-      r0c0:
-        - 13
-        - 1088
-    prescan:
-      rows: 1200
-      cols: 1088
-      # good, reliable cols used for getting row-by-row bias, relative to r0c0
-      col_start: 800
-      col_end: 1000
-      r0c0:
-        - 0
-        - 0
-    parallel_overscan:
-      rows: 163
-      cols: 1056
-      r0c0:
-        - 1037
-        - 1088
-    serial_overscan:
-      rows: 1200
-      cols: 56
-      r0c0:
-        - 0
-        - 2144
-
-# ENG frame:
-ENG:
-  frame_rows: 2200
-  frame_cols: 2200
-  geom:
-    image:
-      rows: 1024
-      cols: 1024
-      r0c0:
-        - 13
-        - 1088
-    prescan:
-      rows: 2200
-      cols: 1088
-      # good, reliable cols used for getting row-by-row bias, relative to r0c0
-      col_start: 800
-      col_end: 1000
-      r0c0:
-        - 0
-        - 0
-    parallel_overscan:
-      rows: 1163
-      cols: 1056
-      r0c0:
-        - 1037
-        - 1088
-    serial_overscan:
-      rows: 2200
-      cols: 56
-      r0c0:
-        - 0
-        - 2144
-  # upper part of ENG frame:
-  eng_upper_rows: 1000
-  eng_upper_cols: 2200
-  eng_geom:
-    eng_parallel_overscan: #cols are in the range of the image cols of SCI half
-      rows: 1000
-      cols: 1056
-      r0c0:
-        - 1200
-        - 1088
-    eng_prescan:
-      rows: 1000
-      cols: 1088
-      r0c0:
-        - 1200
-        - 0
-    eng_serial_overscan:
-      rows: 1000
-      cols: 56
-      r0c0:
-        - 1200
-        - 2144
-
-# time to read out a row in seconds
-rowreadtime_sec: 223.5e-6
-
-# number of EM gain register stages
-Nem: 604
-
-# slice of rows that are used for telemetry
-telem_rows_start: -1
-telem_rows_end: ~
-
-# pixel full well (e-)
-fwc: 90000
-# serial full well (e-) in EM gain register in EXCAM EMCCD
-fwc_em: 100000
-
-# cosmic ray hit rate (hits/m**2/sec)
-X: 5.0e+04
-# pixel area (m**2/pixel)
-a: 1.69e-10
-
-# Maximum allowable EM gain
-gmax: 8000.0
-
-delta_constr: 1.0e-4
-# Overhead time, in seconds, for each collected frame.  Used to compute
-# total wall-clock time for data collection
-overhead: 3
-# Maximum allowed electrons/pixel/frame for photon counting
-pc_ecount_max: 0.1
-# number of read noise standard deviations at which to set the
-# photon-counting threshold
-T_factor: 5
-
-######################
-# These below are ultimately sourced from calibration files
-# dark current (e- / second)
-darke: 1.0e-3
-# clock-induced charge (e- / pixel)
-cic: 1.0e-2
-# read noise (e- / pixel)
-rn: 165.0
-# e- per DN factor (k gain)
-eperdn: 8.7
-# the residual FPN+CIC in the region where bias was calculated (i.e., prescan).
-# In DN.
-bias_offset: 0
