--- corgidrp/util/metadata_eng_conv.yaml
@@ -1,85 +1,0 @@
-#############
-# This file specifies the geometry of the ENG format frame when the
-# ENGINEERING_CONV sequence is used.  The geometry
-# is flipped about the column axis near the middle compared to other sequences,
-# and thus we flip the labels for the serial_overscan and prescan (see below
-# for details).
-#############
-
-# Any regions not specified are assumed to be shielded or unused pixels.
-#
-# Geometry regions are defined as:
-#   rows: number of region rows
-#   cols: number of region columns
-#   r0c0: coordinates (row, col) of the corner closest to (0, 0) of the array
-#
-# For example, in a 5x5 array, the following geometry would look as follows:
-#   rows: 2
-#   cols: 3
-#   r0c0:
-#     - 2
-#     - 1
-#
-# [
-#     [0, 0, 0, 0, 0],
-#     [0, 0, 0, 0, 0],
-#     [0, 1, 1, 1, 0],
-#     [0, 1, 1, 1, 0],
-#     [0, 0, 0, 0, 0]
-# ]
-#
-# where (0, 0) is defined as the top left of this array.
-
-# ENG frame:
-frame_rows: 2200
-frame_cols: 2200
-geom:
-  image: # combined lower and upper
-    rows: 2048
-    cols: 1024
-    r0c0:
-      - 13
-      - 48
-  lower_image:
-    rows: 1024
-    cols: 1024
-    r0c0:
-      - 13
-      - 48
-  upper_image:
-    rows: 1024
-    cols: 1024
-    r0c0:
-      - 1037
-      - 48
-  serial_overscan:
-    rows: 2200
-    cols: 16
-    r0c0:
-      - 0
-      - 0
-  parallel_overscan:
-    rows: 130
-    cols: 1056
-    r0c0:
-      - 2070
-      - 16
-  # This is actually the serial_overscan region, but the code needs to take
-  # the bias from the largest serial non-image region, and the code identifies
-  # this region as the "prescan", so we have the prescan and serial_overscan
-  # names flipped for this reason.
-  prescan:
-    rows: 2200
-    cols: 1128
-    # not sure if these are good in the eng_conv case where the geometry is
-    # flipped relative to the other cases, but these cols would where the
-    # good, reliable cols used for getting row-by-row bias (relative to r0c0)
-    # would be
-    col_start: 1200
-    col_end: 1400
-    r0c0:
-      - 0
-      - 1072
-
-# time to read out a row in seconds
-rowreadtime_sec: 223.5e-6
