--- corgidrp/util/metadata_test.yaml
@@ -1,163 +1,0 @@
-# Used to make a smaller-scale full frame for the purpose of having
-# smaller file sizes for test data for tpump_analysis.
-
-# Any regions not specified are assumed to be shielded or unused pixels.
-#
-# Geometry regions are defined as:
-#   rows: number of region rows
-#   cols: number of region columns
-#   r0c0: coordinates (row, col) of the corner closest to (0, 0) of the array
-#
-# For example, in a 5x5 array, the following geometry would look as follows:
-#   rows: 2
-#   cols: 3
-#   r0c0:
-#     - 2
-#     - 1
-#
-# [
-#     [0, 0, 0, 0, 0],
-#     [0, 0, 0, 0, 0],
-#     [0, 1, 1, 1, 0],
-#     [0, 1, 1, 1, 0],
-#     [0, 0, 0, 0, 0]
-# ]
-#
-# where (0, 0) is defined as the top left of this array.
-
-# total ENG frame:
-eng_frame_rows: 220
-eng_frame_cols: 220
-
-SCI:
-  frame_rows: 120
-  frame_cols: 220
-  geom:
-    image:
-      rows: 104
-      cols: 105
-      r0c0:
-        - 2
-        - 108
-    prescan:
-      rows: 120
-      cols: 108
-      # good, reliable cols used for getting row-by-row bias, relative to r0c0
-      #For such a small test frame, we need enough columns to get an accurate
-      # bias subtraction, so we choose all of them.
-      col_start: 0 #10
-      col_end: 108 #100
-      r0c0:
-        - 0
-        - 0
-    parallel_overscan:
-      rows: 14
-      cols: 107
-      r0c0:
-        - 106
-        - 108
-    serial_overscan:
-      rows: 120
-      cols: 5
-      r0c0:
-        - 0
-        - 215
-
-# ENG frame:
-ENG:
-  frame_rows: 220
-  frame_cols: 220
-  geom:
-    image:
-      rows: 102
-      cols: 102
-      r0c0:
-        - 13
-        - 108
-    prescan:
-      rows: 220
-      cols: 108
-      # good, reliable cols used for getting row-by-row bias, relative to r0c0
-      col_start: 80
-      col_end: 100
-      r0c0:
-        - 0
-        - 0
-    parallel_overscan:
-      rows: 116
-      cols: 105
-      r0c0:
-        - 104
-        - 108
-    serial_overscan:
-      rows: 220
-      cols: 5
-      r0c0:
-        - 0
-        - 215
-  # upper part of ENG frame:
-  eng_upper_rows: 100
-  eng_upper_cols: 220
-  eng_geom:
-    eng_parallel_overscan: #cols are in the range of the image cols of SCI half
-      rows: 100
-      cols: 105
-      r0c0:
-        - 120
-        - 108
-    eng_prescan:
-      rows: 100
-      cols: 108
-      r0c0:
-        - 120
-        - 0
-    eng_serial_overscan:
-      rows: 100
-      cols: 5
-      r0c0:
-        - 0
-        - 215
-
-# time to read out a row in seconds
-rowreadtime_sec: 223.5e-6
-
-# number of EM gain register stages
-Nem: 604
-
-# slice of rows that are used for telemetry
-telem_rows_start: -1
-telem_rows_end: ~
-
-# pixel full well (e-)
-fwc: 90000
-# serial full well (e-) in EM gain register in EXCAM EMCCD
-fwc_em: 100000
-
-# cosmic ray hit rate (hits/m**2/sec)
-X: 5.0e+04
-# pixel area (m**2/pixel)
-a: 1.69e-10
-
-# Maximum allowable EM gain
-gmax: 8000.0
-
-delta_constr: 1.0e-4
-# Overhead time, in seconds, for each collected frame.  Used to compute
-# total wall-clock time for data collection
-overhead: 3
-# Maximum allowed electrons/pixel/frame for photon counting
-pc_ecount_max: 0.1
-# number of read noise standard deviations at which to set the
-# photon-counting threshold
-T_factor: 5
-
-######################
-# These below are ultimately sourced from calibration files
-# dark current (e- / second)
-darke: 1.0e-3
-# clock-induced charge (e- / pixel)
-cic: 1.0e-2
-# read noise (e- / pixel)
-rn: 165.0
-# e- per DN factor (k gain)
-eperdn: 8.7
